workaround for Windows LDAP range queries
Assuming we define

    groupMembershipAttributes: [ member ]

in the conf, the query will work up to 1500 members, else it will return
a member;range value with the first 1500 members.

This commit demonstrate how to support more than 1500 members but no
more than than 3000 by reissuing a query for the next batch of 1500
members.

We hack the result of the LDAP entries so that we have one attribute
named member where we coalesce the various member list.

https://learn.microsoft.com/en-us/previous-versions/windows/desktop/ldap/searching-using-range-retrieval#range-retrieval-results